requirements.txt
fastapi
uvicorn
httpx
liquidpy
aioredis
pytest
pytest-asyncio